requirements.txt
streamlit>=1.36
scikit-learn>=1.3
transformers>=4.41
accelerate>=0.30
sentencepiece>=0.1.99
torch>=2.2
python-dotenv>=1.0
openai>=1.40
faiss-cpu>=1.8.0
sentence-transformers>=2.6.1
